feature(select-component): add screen observer
