+++ challenge-catar Watch App/Views/ContentView.swift
@@ -16,14 +16,6 @@ struct Data: Identifiable {
 
 
 
-class ScreenObserver: ObservableObject {
-
-    @Published var screens: [ScreenAvaiable] = [
-        ScreenAvaiable(id: 0, name: "Relatório Semanal", height: 96),
-        ScreenAvaiable(id: 1, name: "Acompanhe Corrida", height: 96),
-        ScreenAvaiable(id: 2, name: "Acompanhe Sono", height: 96)
-    ]
-}
 
 struct ScreenButton: View {
